requirements.txt
arrow==1.2.3
asgiref==3.5.2
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
Django==4.1
django-browser-reload==1.8.0
django-tailwind==3.5.0
djangorestframework==3.14.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
Pillow==9.5.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.2
urllib3==1.26.15
